+++ runtime/runtime/Input.cpp
@@ -2,14 +2,34 @@
 
 #include "runtime/platform/Platform.h"
 #include "runtime/platform/events/Event.h"
+#include "runtime/platform/events/MouseEvents.h"
 
-#include <bitset>
+using std::vector;
 
-namespace basalt::input {
+namespace basalt {
+
+auto Input::events() const -> const vector<InputEventPtr>& {
+  return mEvents;
+}
+
+auto Input::is_mouse_button_down(const MouseButton button) const -> bool {
+  return mMouseButtonsDown[enum_cast(button)];
+}
+
+void Input::mouse_button_pressed(const MouseButton button) {
+  mEvents.push_back(std::make_unique<MouseButtonPressed>(button));
+  mMouseButtonsDown[enum_cast(button)] = true;
+}
+
+void Input::mouse_button_released(const MouseButton button) {
+  mEvents.push_back(std::make_unique<MouseButtonReleased>(button));
+  mMouseButtonsDown[enum_cast(button)] = false;
+}
+
+namespace input {
 namespace {
 
 std::bitset<KEY_COUNT> sKeyStates;
-std::bitset<MOUSE_BUTTON_COUNT> sButtonStates;
 math::Vec2i32 sMousePos;
 
 void on_key_pressed(const platform::KeyPressedEvent& event) {
@@ -26,23 +46,11 @@ void on_mouse_moved(const platform::MouseMovedEvent& event) {
   sMousePos = event.pos;
 }
 
-void on_button_pressed(const platform::MouseButtonPressedEvent& event) {
-  const auto index = enum_cast(event.button);
-  sButtonStates[index] = true;
-}
-
-void on_button_released(const platform::MouseButtonReleasedEvent& event) {
-  const auto index = enum_cast(event.button);
-  sButtonStates[index] = false;
-}
-
 void platform_event_callback(const platform::Event& event) {
   const platform::EventDispatcher dispatcher(event);
   dispatcher.dispatch<platform::MouseMovedEvent>(&on_mouse_moved);
   dispatcher.dispatch<platform::KeyPressedEvent>(&on_key_pressed);
   dispatcher.dispatch<platform::KeyReleasedEvent>(&on_key_released);
-  dispatcher.dispatch<platform::MouseButtonPressedEvent>(&on_button_pressed);
-  dispatcher.dispatch<platform::MouseButtonReleasedEvent>(&on_button_released);
 }
 
 } // namespace
@@ -60,8 +68,5 @@ auto mouse_pos() -> math::Vec2i32 {
   return sMousePos;
 }
 
-auto is_mouse_button_pressed(const MouseButton button) -> bool {
-  return sButtonStates[enum_cast(button)];
-}
-
-} // namespace basalt::input
+} // namespace input
+} // namespace basalt