+++ runtime/runtime/platform/events/MouseEvents.h
@@ -9,14 +9,6 @@
 
 namespace basalt::platform {
 
-enum class MouseButton : u8 {
-  Unknown = 0,
-  Left, Right, Middle, Button4, Button5,
-};
-
-constexpr uSize MOUSE_BUTTON_COUNT = 6u;
-
-
 struct MouseMovedEvent final : EventTyped<EventType::MouseMoved> {
   math::Vec2i32 pos {};
 
@@ -30,48 +22,6 @@ struct MouseMovedEvent final : EventTyped<EventType::MouseMoved> {
 
   auto operator=(const MouseMovedEvent&) -> MouseMovedEvent& = default;
   auto operator=(MouseMovedEvent&&) -> MouseMovedEvent& = default;
-};
-
-struct MouseButtonPressedEvent final
-  : EventTyped<EventType::MouseButtonPressed> {
-  MouseButton button {MouseButton::Unknown};
-
-  constexpr explicit MouseButtonPressedEvent(const MouseButton button) noexcept
-    : button {button} {
-  }
-
-  constexpr MouseButtonPressedEvent(const MouseButtonPressedEvent&) noexcept
-  = default;
-  constexpr MouseButtonPressedEvent(MouseButtonPressedEvent&&) noexcept
-  = default;
-
-  ~MouseButtonPressedEvent() noexcept = default;
-
-  auto operator=(
-    const MouseButtonPressedEvent&) -> MouseButtonPressedEvent& = default;
-  auto operator=(
-    MouseButtonPressedEvent&&) -> MouseButtonPressedEvent& = default;
-};
-
-struct MouseButtonReleasedEvent final
-  : EventTyped<EventType::MouseButtonReleased> {
-  MouseButton button {MouseButton::Unknown};
-
-  constexpr explicit MouseButtonReleasedEvent(const MouseButton button) noexcept
-    : button {button} {
-  }
-
-  constexpr MouseButtonReleasedEvent(const MouseButtonReleasedEvent&) noexcept
-  = default;
-  constexpr MouseButtonReleasedEvent(MouseButtonReleasedEvent&&) noexcept
-  = default;
-
-  ~MouseButtonReleasedEvent() noexcept = default;
-
-  auto operator=(
-    const MouseButtonReleasedEvent&) -> MouseButtonReleasedEvent& = default;
-  auto operator=(
-    MouseButtonReleasedEvent&&) -> MouseButtonReleasedEvent& = default;
 };
 
 struct MouseWheelScrolledEvent final