+++ runtime/runtime/shared/Types.h
@@ -63,12 +63,6 @@ using f32 = float;
  */
 using f64 = double;
 
-template <typename E, std::enable_if_t<!std::is_enum_v<E>, int> = 0>
-constexpr auto enum_cast(const E) {
-  static_assert(false, "not an enum type");
-  return 0u;
-}
-
 template <typename E, std::enable_if_t<std::is_enum_v<E>, int> = 0>
 constexpr auto enum_cast(const E e) {
   return static_cast<std::underlying_type_t<E>>(e);