+++ runtime/runtime/types.h
@@ -2,6 +2,8 @@
 #ifndef BASALT_RUNTIME_TYPES_H
 #define BASALT_RUNTIME_TYPES_H
 
+#include "Input.h"
+
 #include "shared/Size2D.h"
 #include "shared/Types.h"
 
@@ -10,6 +12,7 @@ namespace basalt {
 struct UpdateContext final {
   f64 deltaTime {};
   Size2Du16 windowSize {Size2Du16::dont_care()};
+  Input input {};
 };
 
 } // namespace basalt